services/pusher/requirements.txt
# Add service-specific dependencies here
confluent-kafka
requests
python-dotenv
